refactor(blockdb): move under database/heightindexdb
