+++ database/heightindexdb/blockdb/database_test.go
