+++ database/heightindexdb/blockdb/errors.go
