+++ database/heightindexdb/blockdb/helpers_test.go
