+++ database/heightindexdb/blockdb/lock_test.go
