+++ database/heightindexdb/blockdb/recovery_test.go
