+++ vms/saevm/cchain/BUILD.bazel
@@ -41,6 +41,7 @@ go_library(
         "//api/metrics",
         "//cache/lru",
         "//database",
+        "//database/heightindexdb/blockdb",
         "//database/prefixdb",
         "//graft/coreth/core/extstate",
         "//graft/coreth/params",
@@ -91,7 +92,6 @@ go_library(
         "//vms/saevm/saedb",
         "//vms/saevm/statesync",
         "//vms/saevm/types",
-        "//x/blockdb",
         "@com_github_ava_labs_libevm//common",
         "@com_github_ava_labs_libevm//common/hexutil",
         "@com_github_ava_labs_libevm//core",
